x86_64

+++ cmuip/central/include/netdevices.h
@@ -101,22 +101,22 @@ static int inline DRV$LOG_FAO(char *c, ...) {
   d.dsc$w_length=strlen(c);
   d.dsc$a_pointer=c;
   va_list args;
-  int argv[18],argc=0;
+  long argv[18],argc=0;
   va_start(args,c);
   while(argc<18) {
-    argv[argc]=va_arg(args,int);
+    argv[argc]=va_arg(args,long);
     argc++;
   }
   va_end(args);
   return IPACP_Interface->ACPI$LOG_FAO(&d,argv[0],argv[1],argv[2],argv[3],argv[4],argv[5],argv[6],argv[7],argv[8],argv[9],argv[10],argv[11],argv[12],argv[13],argv[14],argv[15],argv[16],argv[17]);
 }
 
-static int inline DRV$XLOG_FAO(int i, ...) {
+static int inline DRV$XLOG_FAO(long i, ...) {
   va_list args;
-  int argv[18],argc=0;
+  long argv[18],argc=0;
   va_start(args,i);
   while(argc<18) {
-    argv[argc]=va_arg(args,int);
+    argv[argc]=va_arg(args,long);
     argc++;
   }
   va_end(args);
@@ -131,22 +131,22 @@ static int inline DRV$QL_FAO(char *c, ...) {
   d.dsc$w_length=strlen(c);
   d.dsc$a_pointer=c;
   va_list args;
-  int argv[18],argc=0;
+  long argv[18],argc=0;
   va_start(args,c);
   while(argc<18) {
-    argv[argc]=va_arg(args,int);
+    argv[argc]=va_arg(args,long);
     argc++;
   }
   va_end(args);
   return IPACP_Interface->ACPI$QL_FAO(&d,argv[0],argv[1],argv[2],argv[3],argv[4],argv[5],argv[6],argv[7],argv[8],argv[9],argv[10],argv[11],argv[12],argv[13],argv[14],argv[15],argv[16],argv[17]);
 }
 
-static int inline DRV$XQL_FAO(int i, ...) {
+static int inline DRV$XQL_FAO(long i, ...) {
   va_list args;
-  int argv[18],argc=0;
+  long argv[18],argc=0;
   va_start(args,i);
   while(argc<18) {
-    argv[argc]=va_arg(args,int);
+    argv[argc]=va_arg(args,long);
     argc++;
   }
   va_end(args);
@@ -161,10 +161,10 @@ static int inline DRV$OPR_FAO(char *c, ...) {
   d.dsc$w_length=strlen(c);
   d.dsc$a_pointer=c;
   va_list args;
-  int argv[18],argc=0;
+  long argv[18],argc=0;
   va_start(args,c);
   while(argc<18) {
-    argv[argc]=va_arg(args,int);
+    argv[argc]=va_arg(args,long);
     argc++;
   }
   va_end(args);
@@ -176,10 +176,10 @@ static int inline DRV$ERROR_FAO(char *c, ...) {
   d.dsc$w_length=strlen(c);
   d.dsc$a_pointer=c;
   va_list args;
-  int argv[18],argc=0;
+  long argv[18],argc=0;
   va_start(args,c);
   while(argc<18) {
-    argv[argc]=va_arg(args,int);
+    argv[argc]=va_arg(args,long);
     argc++;
   }
   va_end(args);
@@ -191,10 +191,10 @@ static int inline DRV$WARN_FAO(char *c, ...) {
   d.dsc$w_length=strlen(c);
   d.dsc$a_pointer=c;
   va_list args;
-  int argv[18],argc=0;
+  long argv[18],argc=0;
   va_start(args,c);
   while(argc<18) {
-    argv[argc]=va_arg(args,int);
+    argv[argc]=va_arg(args,long);
     argc++;
   }
   va_end(args);
@@ -206,10 +206,10 @@ static int inline DRV$FATAL_FAO(char *c, ...) {
   d.dsc$w_length=strlen(c);
   d.dsc$a_pointer=c;
   va_list args;
-  int argv[18],argc=0;
+  long argv[18],argc=0;
   va_start(args,c);
   while(argc<18) {
-    argv[argc]=va_arg(args,int);
+    argv[argc]=va_arg(args,long);
     argc++;
   }
   va_end(args);
